+++ application/Espo/Core/Formula/Functions/NumericGroup/MultiplicationType.php
@@ -49,7 +49,7 @@ class MultiplicationType extends \Espo\Core\Formula\Functions\Base
             $part = $this->evaluate($subItem);
 
             if (!is_float($part) && !is_int($part)) {
-                $part = intval($part);
+                $part = floatval($part);
             }
 
             $result *= $part;
